backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic[email]==2.7.1
bcrypt==3.2.2
